+++ plugins/libseqnum.c
@@ -19,6 +19,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #include <gdk/gdk.h>
 #include <gtk/gtk.h>
@@ -155,7 +156,7 @@ PRIVATE void update_label(GtkWidget *label, float val) {
 }
 
 PRIVATE void value_changed_handler(GtkAdjustment *adj, gpointer userdata) {
-  int step = (int) userdata;
+  int step = (intptr_t) userdata;
   Control *c = gtk_object_get_data(GTK_OBJECT(adj), "Control");
   GtkWidget **widgets = c->data;
   Data *data = c->g->data;
@@ -328,7 +329,7 @@ PRIVATE void init_pattern(Control *control) {
 
     gtk_object_set_data(GTK_OBJECT(adj), "Control", control);
     gtk_signal_connect(GTK_OBJECT(adj), "value_changed",
-		       GTK_SIGNAL_FUNC(value_changed_handler), (gpointer) i);
+		       GTK_SIGNAL_FUNC(value_changed_handler), (gpointer) (intptr_t) i);
     gtk_signal_connect(GTK_OBJECT(b), "focus_in_event",
 		       GTK_SIGNAL_FUNC(focus_in_handler), control);
     gtk_box_pack_start(GTK_BOX(hb), b, FALSE, FALSE, 0);
